refactor(type): split combat identity rules
