+++ Source/Portfolio/Type/CActionTypes.cpp
@@ -1,0 +1,7 @@
+#include "Type/CActionTypes.h"
+
+bool FActionDataKey::IsValidMinimal() const
+{
+	return ActionType != EActionType::None
+		&& ActionType != EActionType::Max;
+}
