+++ Source/Portfolio/Type/CReactionTypes.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include "CoreMinimal.h"
+#include "CReactionTypes.generated.h"
+
+UENUM(BlueprintType)
+enum class EReactionType : uint8
+{
+	None = 0,	// Invalid, Unset
+
+	Idle,
+
+	Hit,
+	Dead,
+	BlockHit,
+	Parry,
+	Stagger,
+
+	All,		// Wildcard
+
+	Max,		// Sentinel
+};
+
+UENUM(BlueprintType)
+enum class EReactionNotifyCommand : uint8
+{
+	None = 0,
+
+	Complete,
+
+	Max,
+};
