+++ Source/Portfolio/Type/CWeaponStructure.h
@@ -3,126 +3,13 @@
 #include "CoreMinimal.h"
 #include "Engine/DamageEvents.h"
 #include "DamageEventId.h"
+#include "Type/CWeaponTypes.h"
+#include "Type/CActionTypes.h"
+#include "Type/CReactionTypes.h"
+#include "Type/CExecutionRuleTypes.h"
 #include "Type/CStateTypes.h"
 #include "Type/CHealthTypes.h"
 #include "CWeaponStructure.generated.h"
-
-UENUM(BlueprintType)
-enum class EWeaponType : uint8
-{
-	None = 0,	// Invalid, Unset
-
-	Unarmed,
-	Sword,
-
-	All,		// Wildcard
-
-	Max,		// Sentinel
-};
-
-UENUM(BlueprintType)
-enum class EActionType : uint8
-{
-	None = 0,	// Invalid, Unset
-
-	Idle,
-
-	Equip,
-	Unequip,
-
-	ComboAttack,
-
-	Guard,
-	Dodge,
-
-	All,		// Wildcard
-
-	Max,		// Sentinel
-};
-
-UENUM(BlueprintType)
-enum class EGuardActionPhase : uint8
-{
-	None = 0,
-
-	In,
-	Out,
-	Hold,
-	Hit,
-	Parry,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EReactionType : uint8
-{
-	None = 0,	// Invalid, Unset
-
-	Idle,
-
-	Hit,
-	Dead,
-	BlockHit,
-	Parry,
-	Stagger,
-
-	All,		// Wildcard
-
-	Max,		// Sentinel
-};
-
-UENUM(BlueprintType)
-enum class EActionNotifyCommand : uint8
-{
-	None = 0,
-
-	Complete,
-
-	PushHitContext,
-	ClearHitContext,
-
-	OpenReserveChainWindow,
-	CloseReserveChainWindow,
-	ConsumeChain,
-
-	Equip,
-	Unequip,
-
-	SwitchToGuard,
-	AllowGuardStart,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EReactionNotifyCommand : uint8
-{
-	None = 0,
-
-	Complete,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EActionEventType : uint8
-{
-	None = 0,
-
-	ReserveChainWindowOpened,
-	ReserveChainWindowClosed,
-
-	ActionStarted,
-	ActionCompleted,
-
-	ActionChained,
-
-	ActionInterrupted,
-	ActionIgnored,
-
-	Max,
-};
 
 UENUM(BlueprintType)
 enum class EExecutionDecision : uint8
@@ -168,17 +55,6 @@ enum class EObservableOverlayHandling : uint8
 
 	ClearGuardState,
 	ClearGuardOverlay,
-
-	Max,
-};
-
-UENUM(BlueprintType)
-enum class EExecutionDomain : uint8
-{
-	None = 0,
-
-	Action,
-	Reaction,
 
 	Max,
 };
@@ -235,95 +111,6 @@ public:
 	bool IsValidMinimal() const
 	{
 		return EventType != EObservableOverlayEventType::None && EventType != EObservableOverlayEventType::Max;
-	}
-};
-
-UENUM(BlueprintType)
-enum class EExecutionInterventionTiming : uint8
-{
-	None = 0,
-
-	Always,
-	Window,
-
-	Max,
-};
-
-struct FExecutionParticipant;
-
-USTRUCT(BlueprintType)
-struct FExecutionInterventionParticipantFilter
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere, Category = "Filter")
-	EExecutionDomain Domain = EExecutionDomain::None;
-
-	UPROPERTY(EditAnywhere, Category = "Filter")
-	EActionType ActionType = EActionType::None;
-
-	UPROPERTY(EditAnywhere, Category = "Filter")
-	EReactionType ReactionType = EReactionType::None;
-
-	// INDEX_NONE means any index. Reaction currently ignores Index.
-	UPROPERTY(EditAnywhere, Category = "Filter")
-	int32 Index = INDEX_NONE;
-
-public:
-	bool IsValidMinimal() const;
-
-	bool MatchesAction(EActionType InActionType, int32 InIndex = INDEX_NONE) const;
-	bool MatchesReaction(EReactionType InReactionType) const;
-	bool MatchesParticipant(const FExecutionParticipant& InParticipant) const;
-
-public:
-	bool operator==(const FExecutionInterventionParticipantFilter& InOther) const
-	{
-		return Domain == InOther.Domain
-			&& ActionType == InOther.ActionType
-			&& ReactionType == InOther.ReactionType
-			&& Index == InOther.Index;
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionInterventionWantRule
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere, Category = "Intervention")
-	TArray<FExecutionInterventionParticipantFilter> ParticipantFilters;
-
-public:
-	bool IsValidMinimal() const
-	{
-		return !ParticipantFilters.IsEmpty();
-	}
-};
-
-USTRUCT(BlueprintType)
-struct FExecutionInterventionAllowRule
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere, Category = "Intervention")
-	EExecutionInterventionTiming Timing = EExecutionInterventionTiming::Always;
-
-	UPROPERTY(EditAnywhere, Category = "Intervention")
-	FName WindowKey = NAME_None;
-
-	UPROPERTY(EditAnywhere, Category = "Intervention")
-	TArray<FExecutionInterventionParticipantFilter> ParticipantFilters;
-
-public:
-	bool IsValidMinimal() const
-	{
-		return Timing != EExecutionInterventionTiming::None
-			&& Timing != EExecutionInterventionTiming::Max
-			&& !ParticipantFilters.IsEmpty();
 	}
 };
 
@@ -629,89 +416,6 @@ enum class EActionSFXPlayType : uint8
 };
 
 USTRUCT(BlueprintType)
-struct FActionDataKey
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere, Category = "Key")
-	EActionType ActionType = EActionType::Max;
-
-	UPROPERTY(EditAnywhere, Category = "Key")
-	int32 ActionIndex = INDEX_NONE;
-
-public:
-	bool IsValidMinimal() const;
-
-public:
-	bool operator==(const FActionDataKey& InOther) const
-	{
-		return ActionType == InOther.ActionType
-			&& ActionIndex == InOther.ActionIndex;
-	}
-};
-
-FORCEINLINE uint32 GetTypeHash(const FActionDataKey& InKey)
-{
-	uint32 H = 0;
-
-	H = HashCombine(H, GetTypeHash(static_cast<uint8>(InKey.ActionType)));
-	H = HashCombine(H, GetTypeHash(InKey.ActionIndex));
-
-	return H;
-}
-
-FORCEINLINE int32 GetGuardActionPhaseIndex(EGuardActionPhase InPhase)
-{
-	switch (InPhase)
-	{
-	case EGuardActionPhase::In:
-		return 1;
-
-	case EGuardActionPhase::Out:
-		return 2;
-
-	case EGuardActionPhase::Hold:
-		return 3;
-
-	case EGuardActionPhase::Hit:
-		return 4;
-
-	case EGuardActionPhase::Parry:
-		return 5;
-
-	default:
-		return INDEX_NONE;
-	}
-}
-
-FORCEINLINE EGuardActionPhase ResolveGuardActionPhase(const FActionDataKey& InKey)
-{
-	if (InKey.ActionType != EActionType::Guard) return EGuardActionPhase::None;
-
-	switch (InKey.ActionIndex)
-	{
-	case 1:
-		return EGuardActionPhase::In;
-
-	case 2:
-		return EGuardActionPhase::Out;
-
-	case 3:
-		return EGuardActionPhase::Hold;
-
-	case 4:
-		return EGuardActionPhase::Hit;
-
-	case 5:
-		return EGuardActionPhase::Parry;
-
-	default:
-		return EGuardActionPhase::None;
-	}
-}
-
-USTRUCT(BlueprintType)
 struct FActionData
 {
 	GENERATED_BODY()
@@ -811,35 +515,6 @@ public:
 
 public:
 	bool IsValidMinimal() const;
-};
-
-USTRUCT(BlueprintType)
-struct FWeaponContext
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere)
-	EWeaponType WeaponType = EWeaponType::Max;
-
-public:
-	FWeaponContext() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FActionContext
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(EditAnywhere)
-	EActionType ActionType = EActionType::Max;
-
-	UPROPERTY(EditAnywhere)
-	int32 ActionIndex = INDEX_NONE;
-
-public:
-	FActionContext() = default;
 };
 
 USTRUCT(BlueprintType)
